Make "Baggy" cross-compile
