+++ gen/shared/src/main/scala-2.12/nyaya/gen/Platform.scala
@@ -1,0 +1,8 @@
+package nyaya.gen
+
+object Platform {
+
+  trait BaggyImplicits {
+  }
+
+}
